services/ingestion-service/requirements.txt
fastapi
uvicorn
kafka-python
pydantic
python